Fixed the VLV grammar
git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@1664729 13f79535-47bb-0310-9956-ffa450edef68

+++ ldap/extras/codec/src/main/java/org/apache/directory/api/ldap/extras/controls/vlv_impl/VirtualListViewRequestGrammar.java
@@ -72,7 +72,7 @@ public class VirtualListViewRequestGrammar extends AbstractGrammar<VirtualListVi
                 UniversalTag.INTEGER.getValue(),
                 new StoreAfterCount() );
 
-        super.transitions[VirtualListViewRequestStates.VLV_AFTER_COUNT_STATE.ordinal()][( byte ) VirtualListViewerTags.BY_OFFSET_TAG
+        super.transitions[VirtualListViewRequestStates.VLV_AFTER_COUNT_STATE.ordinal()][VirtualListViewerTags.BY_OFFSET_TAG
             .getValue()] =
             new GrammarTransition<VirtualListViewRequestContainer>(
                 VirtualListViewRequestStates.VLV_AFTER_COUNT_STATE,
@@ -80,7 +80,7 @@ public class VirtualListViewRequestGrammar extends AbstractGrammar<VirtualListVi
                 ( byte ) VirtualListViewerTags.BY_OFFSET_TAG.getValue(),
                 null );
 
-        super.transitions[VirtualListViewRequestStates.VLV_AFTER_COUNT_STATE.ordinal()][( byte ) VirtualListViewerTags.ASSERTION_VALUE_TAG
+        super.transitions[VirtualListViewRequestStates.VLV_AFTER_COUNT_STATE.ordinal()][VirtualListViewerTags.ASSERTION_VALUE_TAG
             .getValue()] =
             new GrammarTransition<VirtualListViewRequestContainer>(
                 VirtualListViewRequestStates.VLV_AFTER_COUNT_STATE,
